script.py
from numbers import Real
from pathlib import Path
import os
from sentence_transformers import SentenceTransformer
from llama_index.core.retrievers import VectorIndexRetriever
from llama_index.embeddings.huggingface import HuggingFaceEmbedding
from llama_index.core import StorageContext, load_index_from_storage, QueryBundle, Settings
from llama_index.vector_stores.hnswlib import HnswlibVectorStore


if __name__ == "__main__":

    print("Hello World!")
