requirements.txt
argparse
numbers
pathlib
os
sentence-transformers
llama_index.core.retrievers
llama_index.embeddings.huggingface
llama_index.core
llama_index.vector_stores.hnswlib
